+++ src/lib/tsr/types.ts
@@ -2,6 +2,7 @@ import * as zod from "zod";
 import { ZodOptionalType, ZodRawShape, ZodType, ZodTypeAny } from "zod";
 import { TSRDefinition } from "./tsr";
 import { PathParams } from "./PathParams";
+import { Route } from "./Route";
 
 export type RouterLocation = "NominalString<RouterLocation>";
 
@@ -35,48 +36,6 @@ export interface RouteRendererProps<Params, RenderResult> {
 export type RouteRenderer<Params, RenderResult> = (
   props: RouteRendererProps<Params, RenderResult>
 ) => RenderResult;
-
-export interface Route<Def extends RouteDefinition = RouteDefinition>
-  extends RouteLocationFactory<InferRouteParams<Def["params"]>> {
-  def: Def;
-
-  /**
-   * Only available for routes in a router.
-   */
-  parent?: AnyRouteLike<Def>;
-
-  render: Def["renderer"];
-
-  parseLocation(location: string): InferRouteParams<Def["params"]> | undefined;
-
-  path<Path extends string>(
-    path: Path,
-    matchOptions?: RouteMatchOptions
-  ): Route<RouteDefinition<Def["tsr"], Path, Def["params"], Def["children"]>>;
-
-  params<ParamsType extends RouteParamsTypeFor<Def["path"]>>(
-    params: ParamsType
-  ): Route<
-    RouteDefinition<Def["tsr"], Def["path"], ParamsType, Def["children"]>
-  >;
-
-  meta(meta: Def["meta"]): Route<Def>;
-
-  renderer(renderer: Def["renderer"]): Route<Def>;
-
-  use(
-    ...additionalMiddlewares: Array<
-      RouteMiddleware<
-        InferRouteParams<Def["params"]>,
-        Def["tsr"]["renderResult"]
-      >
-    >
-  ): Route<Def>;
-
-  children<Children extends RouteMap<Def["tsr"]>>(
-    children: Children
-  ): Route<RouteDefinition<Def["tsr"], Def["path"], Def["params"], Children>>;
-}
 
 export type RouteMap<TSRDef extends TSRDefinition = TSRDefinition> = Record<
   string,
